my_agent/requirements.txt
streamlit
openai
requests
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-core
docx2txt
faiss-cpu
dashscope
